requirements.txt
pyzmq==16.0.2
requests==2.18.4